Crate expert_sys

Modules

command
The command module is a User Interface.

ops
The ops module is a collection of operators like: Axiom, And, Or, Xor, Not, [...].

Macros

format_exp!
The format_exp macro returns the String of all dependencies from a expression.

parse_index!
The parse_index macro returns a formated Index or Nothing.

Structs

Axiom
The Axiom structure is a primitive.

Set
The Set structure is all alphabet axioms.

Solver
The Solver structure is all alphabet expression tree.

Traits

Exp
The Exp Trait is a expression's implemention.